refactor: reusable todolist component
